fix: add typings and move function entry
